requirements.txt
flask==2.1.0
werkzeug==2.1.0
mysql-connector-python==8.0.26
python-dotenv==0.19.0
